At least use long for return type of read_file().
int is not acceptable as a return type for anything regarding sizes.
long is dubious as well, but better (64-bit on sane ABIs and the return
type of ftell()).

+++ file_path.h
@@ -26,7 +26,7 @@
 extern "C" {
 #endif
 
-int read_file(const char *path, void **buf);
+long read_file(const char *path, void **buf);
 bool read_file_string(const char *path, char **buf);
 bool write_file(const char *path, const void *buf, size_t size);
